requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
sec_api==1.0.18
yahoo_fin==0.8.9.1
yfinance==0.2.40
